integrated css for add/edit item-images
