+++ app/templates/components/favourite-star.hbs
@@ -1,3 +1,3 @@
 {{#if showStar }}
-  <img src='assets/images/star_circle.png' class='favourite'></img>
+  <span class="icon-key-photo"></span>
 {{/if}}
