+++ app/templates/items/new.hbs
@@ -1,3 +1,12 @@
-<h2>{{t "items.new.add_photos"}}</h2>
+<section class="main-section">
 
-{{upload-images item=this next="next" back="back" updateItem="updateItem"}}
+  <div class="row">
+    <div class="small-12 columns with-btm">
+
+      <h2 class="text-center photo-title">{{t "items.new.add_photos"}}</h2>
+
+      {{upload-images item=this next="next" back="back" updateItem="updateItem"}}
+
+    </div>
+  </div>
+</section>
